+++ biobb_mem/lipyphilic_biobb/__init__.py
@@ -1,2 +1,2 @@
 name = "lipyphilic_biobb"
-__all__ = ["lpp_assign_leaflets", "lpp_zpositions"]
+__all__ = ["lpp_assign_leaflets", "lpp_flip_flop", "lpp_zpositions"]